Add Windows cross-compilation test to daily build
This PR brings back the test of Windows cross-compilation to the daily build; i.e., building enclave binaries in a linux container and testing the binaries on the Windows host.
This allows the daily build to catch bugs that cause such workflow to fail.

Note that the PR only adds two of the configurations, which should be sufficient to test the workflow. The full list of configurations is tested by nightly build.

Signed-off-by: Ming-Wei Shih <mishih@microsoft.com>