Fix tcs exhaustion test to predictably terminate
Previously each thread would keep making ocalls to check if it can exit.
This could cause the thread to hog the CPU, preventing other threads
from making progress so that the termination condition for the test
is reached.

Wait on a condition variable instead. This allows other threads to run.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>