Retain ECALL worker thread's tcs binding.
Instead of returning when spin count threshold has been reached,
make an ocall to sleep in the host.
This ensures that the tcs is still bound to the host thread.
When the messages arrive, the worker thread can go back into the
enclave in the same tcs.
Without retaining the tcs binding, there may not be a tcs available
for the worker thread to go back into the enclave.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>