Fix inconsistent oesign --config-file requirements
Prior to 8c89a3c, oesign treated --config-file as an optional parameter since
it was expected that the .oeinfo block might already exist and not require
redefining at sign time. Instead, there exists `_check_for_missing_options()`
to verify that a config file contained all the necessary options if the
enclave is loaded but the .oeinfo section is not found. That condition is not
valid anymore as changes to oesign and the enclave loader over time has made
it so that `oe_load_elf_enclave_image()` does not even succeed if it cannot
locate the .oeinfo RVA i.e. the absence of .oeinfo cannot be determined as a
step after loading the enclave since that will always fail first.

This change removes the mandatory check for the --config-file option and the
unused `_check_for_missing_options()` function. oesign requires that enclaves
must already have the .oeinfo section to be signed.

Signed-off-by: Simon Leet <simon.leet@microsoft.com>