Use platform's OCaml to build edger8r if available.
This restores the original way in which oeedger8r was built
on both platforms, if ocaml compiler is already available.
If not, it lets esy build the ocaml compiler and then build
oeedger8r.

Note: ocaml needs to be added back to prereqs on both platforms
to ensure a pain free/fast build of oeedger8r.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>