Reduce committed SGX enclave size on Windows
oe_sgx_create_enclave() currently only takes the enclave size as defined by its
ELRANGE, which can be significantly larger than the actual EPC memory that
needs to be committed to load and launch the enclave.

This change updates oe_sgx_create_enclave() to take both an initial commit size
in addition to the enclave ELRANGE size, which should expand the range of
enclave sizes that can successfully load on Windows (see previous issues such
as #2394 and #2397).

Signed-off-by: Simon Leet <simon.leet@microsoft.com>