Usability oriented pointer checks.
When an enclave pointer is passed by the host back into the enclave
either as an in or in/out parameter detect it an raise OE_INVALID_PARAMETER.

Note: OE SDK marshalling strategy does not require these pointers to be passed in.
It is not a security issue. The pointers are now being passed in just for doing
these checks.

Note: Similar usability checks for out parameters are not implemented since
that may require changes to oeedger8r.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>