Reduce maximum threads to 8 for snmalloc sample.
This ensures that the sample works on all known ACC VM configurations.

The total heap requirement for the sample is estimated to be:
   max-num-threads * (snmalloc cost per thread + max-num-objects-per-thread * max-object-size)
   = 16 * (256 KB + 16 + 16KB)
   = 16 * 512KB
   = 16 * 512KB / 4096 pages
   = 2048 pages

Eventhough the memory (heap) requirement is 2048 pages, snmalloc raises
"Critical error: Out-of-memory during initialisation." even when the number of heap pages
are set to 4096 or 8192.

This PR worksaround the above behavior.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>