Add mechanism to pass segfault address to OE exception handler
This commit takes the address that caused a segfault, as reported in the
host signal handler, and passes it to the in-enclave OE exception
handler. This is needed so that an application signal handler can
obtain the faulting address inside the enclave.

The current implementation should be generalised to handle arbitrary
synchronous signals correctly. Note that SGX2 permits the exception
information to be obtained in-enclave, making this untrusted mechanism
unnecessary.

Signed-off-by: Peter Pietzuch <prp@doc.ic.ac.uk>