Update VisualStudioWindows.md
i've updated the LLVM download link with version 10. I suppose it would be beneficial to get the latest one to get the updates and fixes. 
In my usecase, i need to use nlohmann/json. version 7 doesn't work (json::dump() ).  However when i upgraded to v10, it started working (as it is on the list of tested clang/llvm versions)