Add CCC Technical Charter
The policy for all CCC projects
(found at
https://github.com/confidential-computing/governance/blob/master/project-progression-policy.md)
says under Project Acceptance Process:
"The proposal document will be finalized and a project charter prepared.
This charter document must be included in the project's main
repository."

This PR adds the approved charter in the project's main repository as
required.  I will note that the charter is not correct with respect to
OE governance, but that requires a CCC process to amend.  In the
meantime, we need to include this as is to be compliant.  (Today was
the first time I saw this charter, future charters will be visible to
the TAC.)

Signed-off-by: Dave Thaler <dthaler@ntdev.microsoft.com>