Figure out enclave_base_address without help from loader.
Initialize _enclave_rva to its own address:
   OE_EXPORT volatile void* _enclave_rva = &_enclave_rva;

Due to this an X64_RELATIVE relocation entry will be created for _enclave_rva.
Additionally, in the image and in the enclave before relocation,
_enclave_rva will have its own RVA value. This is asserted by the loader.

Thus, before relocation,
   (uint8_t*) &_enclave_rva - (uint64_t) _enclave_rva
will give the enclave's base address which is stored in another variable for use.
After relocation, the above expression will evaluate to zero.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>