Getting Started Guide for those consuming the OE SDK
Previously the main README file implied the Getting Started Guide was
only for those contributing to the OE SDK itself, not those that just
want to consume it.  This is not the case, as the guide (also) already
covered using it and running the samples, and contains information that
anyone consuming it needs to know (e.g., installing prereqs) and was
not documented anywhere else.

In theory we could split/duplicate the Getting Started Guide and have
separate guides for those consuming vs contributing to the OE SDK, but
this PR just makes the minimal changes needed to unblock those who
want to consume the SDK.  SIG-Documentation can discuss whether to
later split and have two guides or not, but that is outside the scope of
this PR.

Signed-off-by: Dave Thaler <dthaler@ntdev.microsoft.com>