README files for SIGs
A few SIGs already had README files and others did not.  This PR
makes all SIGs have consistent README files.

PR #3412 updated the OWNERS files with chairs, but did not make
corresponding changes to the README files which still said "TBD".
This PR addresses that inconsistency.

Names simply match those already in the OWNERS files in the same
directories. Any changes to names should be in a separate PR, since
those are up to SIG chairs.

Any other content changes to the README files would be up to the
relevant SIGs.

Signed-off-by: Dave Thaler <dthaler@ntdev.microsoft.com>