devex: vscode: Use WebPack
Thus far, the VS Code extension has been relying on its Node dependencies
being included in the VSIX package as-is. This is not recommended as it
increases the size of the resulting package. The recommended way of
packaging dependencies is WebPack. Additionally, the VSIX packaging tool,
vsce, explicitly ignores the Node dependencies. Hence, as it was, the
extension ceased to build.

This patch enables the proper use of WebPack for the extension.

Signed-off-by: Hernan Gatta <hegatta@microsoft.com>