Fix for codeql alert - Query ID:cpp/wrong-number-format-arguments
The IF block in OE_RAISE_MSG macro is causing ~160 security errors in CodeQL scans for "Too few arguments to formatting functions".
gcc/clang handles ##VA_ARGS internally. When ##VA_ARGS is empty the comma before the ‘##’ will be deleted.
This IF block is dead code even when in empty arguments scenario.

Signed-off-by: mrragava <mrragava@gmail.com>