compiler-rt for enclaves
GCC/clang generate calls to functions defined in "compiler-rt" when they encounter certain constructs
that don't have a direct mapping to a processor instruction.
See https://gcc.gnu.org/onlinedocs/gcc-4.9.4/gccint/Integer-library-routines.html
These functions are normally available in libgcc.a library.

Clang implements these functions in a "compiler-rt" library.
See: https://compiler-rt.llvm.org/
The files in this PR provide the builtins as described in the link above.

The code has been borrowed from Intel SDK.
https://github.com/intel/linux-sgx/tree/master/sdk/compiler-rt

Compiler-rt is one of the core libraries and when compiling a program normally,
the compiler would automatically add this library to the linker line.

The mirror the same behavior, compiler-rt is built as an object library which
is then added to oecore. This ensures that compiler-rt routines appear at the
appropriate position in the linker line - the last library. By adding the objects
to oecore, there is no change to the user experience, and thus this is not a
breaking change.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>