allow enc props to be overridden for enc create
enclaves can have default enclave properties. during sign they can be
changed as part of the signing config file.

If we are using an unsigned enclave that can be used in different
scenarios we may need to change those default properties without
requiring it to be signed. A new setting has been added so it can be
passed through to oe_enclave_create such that the properties can be
overridden.

This setting is only really useful during development where an unsigned
enclave binary may not be practical.

Signed-off-by: Paul Allen <paul.c.allen@microsoft.com>