Analyze missing symbols
Whenever a module is loaded within an enclave, oegdb will
keep track of the symbols exported by the module as well
as the symbols imported by the module.
To do this, oegdb invoked `objdump -T module-elf-path`.
Those symbols that have an *UND* annotation are undefined,
and hence imported. All other symbols are exported.

The imported symbols for a loaded module are searched for
in already loaded modules. If not found, it indicates a potential
runtime failure. The missing symbol is logged in red.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>