Fix thread local section alignment issue.
Earlier, we used to choose the larger of the two alignments to align both the sections.
This is incorrect. For example, if tbss has a smaller alignment than tdata, then using
the larger alignment value will increase the size of the tdata section. This will inturn
cause a shift in the start of the tdata section which immediately precedes tbss section.
The tdata template will be copied to the region starting at the incorrect tdata start.

The fix is to
- First compute the aligned start of the tbss section.
- Then subract the tdata size from tbss start and then align it to data alignment to obtain tdata start.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>