Fix thread-local alignment issue with lld linker
When lld is used to link enclaves (typically in Windows), there is a mismatch
between the alignment values in tdata section and PT.TLS program header.
The compiler generates code based on the alignment value in PT.TLS program header,
whereas OE relies on tdata's alignment. This results in a mismatch between
OE's thread-local template instantation and compiler's expectation.

This issue does not manifest with GNU ld which is default on Linux.
On Windows, where lld is used for enclave, this issue manifests only for
certain combinations of alignment values of tdata and tbss sections, and
for specific enclave binary code. One cannot predict whether given a tbss value and
tdata value, the issue would arise.

The fix is to use PT.TLS alignment for tdata section.

Long term, OE needs to follow the strategy of relying only on PT.TLS program header
as is done by MUSL.

This PR also adds tests locking down correct behavior for various alignment combinations
of tdata and tbss. For, many of the combinations being tried out, the issue is seen
to occur and the fix is validated.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>