SGX Debugging: Add ocall to log backtrace. Used in oe_abort
It is a long standing issue that oe_abort does not reveal enough information
about why an enclave is aborting. oe_abort issues are typically found by
running in the debugger. Running within the debugger is not possible in
certain workflows.

To solve this, a new OCALL oe_sgx_log_backtrace is introduced which, when
invoked from within the enclave, will log the current enclave backtrace.
oe_abort makes use of this call to log the backtrace prior to marking the
enclave as aborted. This OCALL does not require enclave memory allocation
and thus oe_abort due to out of memory errors should also be logged
(needs to be validated).

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>