Fix: Ubuntu 20.04 getting started driver URL
On step 2 (Install the Intel SGX DCAP Driver) of the getting started documentation for Ubuntu 20.04, the URL to the Intel SGX DCAP driver no longer worked. I have updated the documentation with the URL that is stored in variable file `focal.yml` when using the "linux\sgx" Ansible role's "sgx-driver" task.

Signed-off-by: Thomas Van Laere <4990509+ThomVanL@users.noreply.github.com>